requirements.txt
flask
openpyxl
google-generativeai
pandas
PyPDF2
